+++ website-backend/admin/component-blocks/do-and-avoid.tsx
@@ -79,7 +79,9 @@ export const doAndAvoid = component({
 					doImage={typeof props.fields.doImage.value === 'string' ? props.fields.doImage.value : ''}
 					doText={props.fields.doText.element}
 					dontAlt={props.fields.dontAlt.value}
-					dontImage={typeof props.fields.dontImage.value === 'string' ? props.fields.dontImage.value : ''}
+					dontImage={
+						typeof props.fields.dontImage.value === 'string' ? props.fields.dontImage.value : ''
+					}
 					dontText={props.fields.dontText.element}
 				/>
 			</GEL>
@@ -137,6 +139,7 @@ const Image = (props: ImgHTMLAttributes<HTMLImageElement>) => {
 	const { SPACING } = useBrand();
 	return (
 		<img
+			alt=""
 			css={mq({
 				width: '100%',
 				height: 'auto',