+++ components/header/CHANGELOG.md
@@ -1,6 +1,23 @@
 # @westpac/header
 
+## 1.1.0
+
+### Minor Changes
+
+- 39e90efa: Updated components to export TypeScript interfaces using existing prop types.
+
+### Patch Changes
+
+- 8d3f5b3d: ESLint implementation
+- Updated dependencies [39e90efa]
+- Updated dependencies [8d3f5b3d]
+  - @westpac/a11y@2.1.0
+  - @westpac/button@2.1.0
+  - @westpac/core@3.1.0
+  - @westpac/icon@2.0.1
+
 ## 1.0.0
+
 ### Major Changes
 
 - 45eef270: Update to use correct icons and design fixes